Apply suggested changes
Apply suggested changes
