+++ src/workflows/queueStore.ts
@@ -9,22 +9,34 @@ import { readManifest, writeManifest, withLock } from '../persistence/runDirecto
 import { createLogger, type StructuredLogger, LogLevel } from '../telemetry/logger';
 
 // V2 WAL Components
-import { ensureV2Format } from './queueMigration.js';
-import {
-  hydrateIndex,
-  getCounts,
-  addTask as addTaskToIndex,
-} from './queueMemoryIndex.js';
-import {
-  appendOperationsBatch,
-} from './queueOperationsLog.js';
+import { getCounts, addTask as addTaskToIndex } from './queueMemoryIndex.js';
+import { appendOperationsBatch } from './queueOperationsLog.js';
 import { compact } from './queueCompactionEngine.js';
-import {
-  QUEUE_FILE, QUEUE_MANIFEST_FILE, QUEUE_SNAPSHOT_FILE,
-  buildDependencyGraph, toExecutionTask, toExecutionTaskData,
-  type QueueIndexState, type QueueOperation,
-  type QueueManifest, type QueueSnapshot, type QueueOperationResult,
+import { QUEUE_FILE, QUEUE_MANIFEST_FILE, QUEUE_SNAPSHOT_FILE } from './queueConstants.js';
+import type {
+  QueueOperation,
+  QueueManifest,
+  QueueSnapshot,
+  QueueOperationResult,
 } from './queueTypes.js';
+
+// V2 Cache and helpers (shared with other queue modules to avoid circular dependencies)
+import {
+  getV2IndexCache,
+  invalidateV2Cache,
+  buildDependencyGraph,
+  toExecutionTask,
+  toExecutionTaskData,
+} from './queueCache.js';
+
+// Re-export for backward compatibility
+export {
+  getV2IndexCache,
+  invalidateV2Cache,
+  buildDependencyGraph,
+  toExecutionTask,
+  toExecutionTaskData,
+} from './queueCache.js';
 
 /**
  * Queue Store
@@ -37,9 +49,13 @@ import {
 
 // --- Types ---
 
-// Re-export types and pure helpers from queueTypes for backward compatibility
-export type { QueueManifest, QueueSnapshot, QueueOperationResult, QueueValidationResult } from './queueTypes.js';
-export { buildDependencyGraph, toExecutionTask, toExecutionTaskData } from './queueTypes.js';
+// Re-export types for backward compatibility
+export type {
+  QueueManifest,
+  QueueSnapshot,
+  QueueOperationResult,
+  QueueValidationResult,
+} from './queueTypes.js';
 
 // Module-level logger for queue store operations
 const logger: StructuredLogger = createLogger({
@@ -47,78 +63,6 @@ const logger: StructuredLogger = createLogger({
   minLevel: LogLevel.DEBUG,
   mirrorToStderr: true,
 });
-
-// --- V2 Index State Management ---
-
-/** V2 Index cache entry with state and metadata. */
-interface V2IndexCache {
-  /** Hydrated index state */
-  state: QueueIndexState;
-  /** Queue directory path */
-  queueDir: string;
-  /** Feature ID for this queue */
-  featureId: string;
-  /** Last hydration timestamp */
-  hydratedAt: number;
-  /** Whether V2 format has been verified */
-  migrationChecked: boolean;
-}
-
-/** V2 index state cache, keyed by runDir. */
-const v2IndexCache = new Map<string, V2IndexCache>();
-
-/**
- * Get or create V2 index cache entry.
- * Hydrates index state from V2 WAL format, auto-migrating V1 queues.
- */
-export async function getV2IndexCache(runDir: string): Promise<V2IndexCache> {
-  const manifest = await readManifest(runDir);
-  const queueDir = path.join(runDir, manifest.queue.queue_dir);
-  const featureId = manifest.feature_id;
-
-  const existing = v2IndexCache.get(runDir);
-
-  // Return existing cache if available and fresh
-  if (existing && existing.queueDir === queueDir && existing.migrationChecked) {
-    return existing;
-  }
-
-  // Ensure V2 format (auto-migrate from V1 if needed)
-  const migrationResult = await ensureV2Format(queueDir, featureId);
-  if (migrationResult.result && !migrationResult.result.success) {
-    throw new Error(
-      `Queue migration failed: ${migrationResult.result.error ?? 'Unknown error'}`
-    );
-  }
-  if (migrationResult.migrated && migrationResult.result) {
-    logger.warn('⚠️ V1 queue format detected - auto-migrated to V2', {
-      tasks_converted: migrationResult.result.tasksConverted,
-      backup_path: migrationResult.result.backupPath,
-      from_version: migrationResult.result.fromVersion,
-      to_version: migrationResult.result.toVersion,
-      message: "Run 'ai-feature queue verify' to confirm migration integrity",
-    });
-  }
-
-  // Hydrate index from snapshot + WAL
-  const state = await hydrateIndex(queueDir);
-
-  const cache: V2IndexCache = {
-    state,
-    queueDir,
-    featureId,
-    hydratedAt: Date.now(),
-    migrationChecked: true,
-  };
-
-  v2IndexCache.set(runDir, cache);
-  return cache;
-}
-
-/** Invalidate V2 cache for a run directory. Forces re-hydration on next access. */
-export function invalidateV2Cache(runDir: string): void {
-  v2IndexCache.delete(runDir);
-}
 
 // --- Queue Initialization ---
